requirements.txt
fastapi
uvicorn
pydantic-settings
PyPDF2
faiss-cpu
sentence-transformers
openai
